Use CT for leader tests, let agent monitor nodes, incr leader test
Tries to address a failing case found by Garret Smith:

This way works as expected (no deadlock):
* start a, start locks, start locks_leader process
* start b, start locks, start locks_leader process, connect a to b
* start c, connect to a

This way deadlocks:
* start a, start locks, start locks_leader process
* start b, connect a to b, start locks, start locks_leader process
-- at this point, everything is fine --
* start c, connect to a, locks_leader process on 'a' is deadlocked

The problem seemed to be a race condition between monitor_nodes()
detection in the leader and agent, causing node a to lose track of
the lock on b. This was provisionally fixed by letting the agent
monitor the nodes for the leader, notifying it only when the locks
server is running/terminated on a given node.

It seems probable that a race condition still exists in the agent,
which can be triggered if a lock request arrives just before the
agent is told that a new node is running. However, having the leader
monitor the nodes via the agent reduces the risk of falling into this
hole. It is also possible that it can only happen when using continuous
locking patterns like the leader does.